+++ NEWS.md
@@ -1,5 +1,10 @@
 # In progress
 
+# tshfr 0.0.1
+
+* Added `ts_pdf_document()` and a "LaTeX Article" R Markdown
+  template to simplify PDF creation.
+
 # tshfr 0.0.0.9001
 
 * Added `setdiff3()` and `table_na()` functions.